Fix: Remove padding from external links in WebView
The rule for 'a.external' included 'padding-right: 15px' to make
space for a background icon. This created large, unwanted horizontal
spacing in the WebView when the icon failed to render.

This fix comments out the properties related to the background icon,
primarily removing the padding and resolving the text spacing issue.